LangChain_Deployment_and_Monitoring/LangServe/Langserve_FastAPI_Template/requirements.txt
langchain
python-dotenv
langchain-groq
langchain_openai
langchain-core
fastapi
uvicorn
langserve
sse_starlette